pyproject.toml
[project]
name = "xretrieval"
version = "0.1.1"
description = "Retrieve and Evaluate with X(any) models"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
keywords = ["retrieval", "evaluation", "machine-learning", "multi-modal"]
authors = [
    { name = "Dickson Neoh", email = "dickson.neoh@gmail.com" }
]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torchmetrics>=1.6.0",
    "Pillow>=11.0.0",
    "faiss-cpu>=1.9.0",
    "matplotlib>=3.9.0",
    "loguru>=0.7.0",
    "rich>=13.9.0",
    "tqdm>=4.67.0",
    "pandas>=2.1.0",
    "transformers>=4.46.0",
    "sentence-transformers>=3.3.0",
    "timm>=1.0.0",
    "accelerate>=1.1.0",
    "bm25s>=0.2.5",
    "pystemmer>=2.2.0.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=22.0",
    "isort>=5.0",
    "flake8>=4.0",
]

docs = [
    "ipykernel",
    "livereload",
    "nbconvert",
    "nbformat",
    "sphinx",
    "watchdog",
    "mkdocs",
    "mkdocs-git-revision-date-plugin",
    "mkdocs-git-revision-date-localized-plugin",
    "mkdocs-jupyter>=0.24.0",
    "mkdocs-material>=9.1.3",
    "mkdocs-pdf-export-plugin",
    "mkdocstrings",
    "mkdocstrings-crystal",
    "mkdocstrings-python-legacy",
    "pygments",
    "pymdown-extensions",
]



[project.urls]
"Homepage" = "https://github.com/dnth/x.retrieval"
"Bug Tracker" = "https://github.com/dnth/x.retrieval/issues"

[tool.black]
line-length = 88
target-version = ["py37"]

[tool.isort]
profile = "black"
multi-line-output = 3 
[tool.pixi.project]
channels = ["conda-forge"]
platforms = ["linux-64"]

[tool.bumpversion]
current_version = "0.1.1"
commit = true
tag = true

[[tool.bumpversion.files]]
filename = "pyproject.toml"
search = 'version = "{current_version}"'
replace = 'version = "{new_version}"'

[[tool.bumpversion.files]]
filename = "xretrieval/__init__.py"
search = '__version__ = "{current_version}"'
replace = '__version__ = "{new_version}"'

[tool.hatch.build]
include = [
    "xretrieval/**/*.py",
]
exclude = [
    "tests/*",
    "nbs/*",
]